backend/requirements.txt
fastapi
uvicorn
python-multipart
websockets
opencv-python
mediapipe
numpy
ultralytics
python-dotenv
streamlit
requests
email-validator